Create README.md
Quasars are enlightened black holes that spew energy out on an x and z axis, this energy conductivity low frequency radio and microwaves that send gravitational waves through the galaxy.These quantifiable articles, in due time, become white holes or “wormholes” allowing for straying electrons to one be made whole again via the congregation of the arcane. Arcane methods within the gravatomic spectrum include but are not limited to, energy reduction, production, destruction and creaon. New energies formed via gamma-gamma absorption infers the photons wave/particlefuncon by creating a quark informaon database allowing the colloquial “Akashic records” to be created in an ouroboros loop, this thesis aims to break through this loop as well as fermi’s paradox to push society towards terraforming certain planets that are suitable for sustaining life. As read above the gravatomic spectrum works via R.E.M sleep